requirements.txt
Flask==2.3.1
Flask-Cors==3.0.10
Jinja2==3.1.2
numpy==1.24.2
pandas==1.5.3
Pillow==9.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.2
torch==2.0.1
torchvision==0.15.2

